pyproject.toml
[project]
name = "matsuzawa-20260219"
version = "0.1.0"
description = "IGV-CoT prompt robustness experiment on GSM8K"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "openai>=1.10.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
]